+++ eslint.config.mjs
@@ -6,10 +6,29 @@ import typescript from '@metamask/eslint-config-typescript';
 const NODE_LTS_VERSION = 22;
 
 /**
+ * Arguments to the `no-restricted` syntax rule that advises use of
+ * `${Controller}:stateChanged` instead of `:stateChange`.
+ */
+const NO_CONTROLLER_STATE_CHANGE_SELECTOR_OBJECTS = [
+  {
+    selector:
+      'CallExpression[callee.property.name="subscribe"] > Literal[value=/^.+:stateChange$/]',
+    message:
+      "Subscribing to ':stateChange' events is deprecated. Use ':stateChanged' instead.",
+  },
+  {
+    selector:
+      'CallExpression[callee.property.name="delegate"] Property[key.name="events"] ArrayExpression > Literal[value=/^.+:stateChange$/]',
+    message:
+      "Delegating ':stateChange' events is deprecated. Use ':stateChanged' instead.",
+  },
+];
+
+/**
  * Arguments to the `no-restricted-syntax` rule that prevents messsenger actions
  * from being called in constructors.
  */
-const NO_MESSENGER_ACTIONS_IN_CONSTRUCTORS_RULES = [
+const NO_MESSENGER_ACTIONS_IN_CONSTRUCTORS_SELECTOR_OBJECTS = [
   {
     selector:
       'MethodDefinition[kind="constructor"] CallExpression[callee.type="MemberExpression"][callee.property.name="call"][callee.object.type="MemberExpression"][callee.object.object.type="ThisExpression"][callee.object.property.name="messenger"]',
@@ -126,6 +145,17 @@ const config = createConfig([
       // do not work very well.
       'jsdoc/check-tag-names': 'off',
       'jsdoc/require-jsdoc': 'off',
+
+      // Add custom rule for deprecating `${Controller}:stateChange` in favor of
+      // `:stateChanged`.
+      'no-restricted-syntax': [
+        'error',
+        ...collectExistingRuleOptions('no-restricted-syntax', [
+          base,
+          typescript,
+        ]),
+        ...NO_CONTROLLER_STATE_CHANGE_SELECTOR_OBJECTS,
+      ],
     },
   },
   {
@@ -208,7 +238,8 @@ const config = createConfig([
           base,
           typescript,
         ]),
-        ...NO_MESSENGER_ACTIONS_IN_CONSTRUCTORS_RULES,
+        ...NO_CONTROLLER_STATE_CHANGE_SELECTOR_OBJECTS,
+        ...NO_MESSENGER_ACTIONS_IN_CONSTRUCTORS_SELECTOR_OBJECTS,
       ],
     },
   },
@@ -223,7 +254,8 @@ const config = createConfig([
           base,
           typescript,
         ]),
-        ...NO_MESSENGER_ACTIONS_IN_CONSTRUCTORS_RULES,
+        ...NO_CONTROLLER_STATE_CHANGE_SELECTOR_OBJECTS,
+        ...NO_MESSENGER_ACTIONS_IN_CONSTRUCTORS_SELECTOR_OBJECTS,
         {
           selector:
             'ExportNamedDeclaration > ExportSpecifier[local.name=/AllowedActions$/]',
@@ -257,6 +289,12 @@ const config = createConfig([
   },
   {
     files: ['packages/messenger/src/generate-action-types/**/*.{js,ts}'],
+    rules: {
+      'import-x/no-nodejs-modules': 'off',
+    },
+  },
+  {
+    files: ['packages/messenger-cli/src/**/*.{js,ts}'],
     rules: {
       'import-x/no-nodejs-modules': 'off',
     },